Improved "Left-right borders" option Changed default to 15, to match VSFilter
git-svn-id: https://ffdshow-tryout.svn.sourceforge.net/svnroot/ffdshow-tryout/trunk@3432 3b938f2f-1a1a-0410-8054-a526ea5ff92c

+++ src/subtitles/TfontSettings.h
@@ -16,7 +16,7 @@ public:
  unsigned int getSize(unsigned int refResY,unsigned int clipdy) const
   {
    if (autosize && clipdy && refResY)
-      return (sizeA * clipdy / refResY) + 0.5;
+      return (int)((sizeA * clipdy / refResY) + 0.5);
    else
       return sizeP;
   }
